fix: add star dashboard functionality
