+++ src/pages/view/TitleBar/TitleBar.js
@@ -1,10 +1,18 @@
+import {
+    useDataEngine,
+    useAlert,
+} from '@dhis2/app-runtime'
+import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { connect } from 'react-redux'
+import { acSetDashboardStarred } from '../../../actions/dashboards.js'
+import { sGetDashboardStarred } from '../../../reducers/dashboards.js'
 import { sGetSelected } from '../../../reducers/selected.js'
 import { sGetShowDescription } from '../../../reducers/showDescription.js'
 import FilterBar from '../FilterBar/FilterBar.js'
 import ActionsBar from './ActionsBar.js'
+import { apiStarDashboard } from './apiStarDashboard.js'
 import Description from './Description.js'
 import LastUpdatedTag from './LastUpdatedTag.js'
 import StarDashboardButton from './StarDashboardButton.js'
@@ -15,7 +23,32 @@ const ViewTitleBar = ({
     displayName,
     displayDescription,
     showDescription,
+    starred,
+    setDashboardStarred,
+
 }) => {
+
+    const dataEngine = useDataEngine()
+    const { show } = useAlert(
+        ({ msg }) => msg,
+        ({ isCritical }) =>
+            isCritical ? { critical: true } : { warning: true }
+    )
+
+
+    const onToggleStarredDashboard = (cb) =>
+        apiStarDashboard(dataEngine, id, !starred)
+            .then(() => {
+                setDashboardStarred(id, !starred)
+                cb && cb()
+            })
+            .catch(() => {
+                const msg = starred
+                    ? i18n.t('Failed to unstar the dashboard')
+                    : i18n.t('Failed to star the dashboard')
+                show({ msg, isCritical: false })
+            })
+
     return (
         <>
             <div className={classes.container}>
@@ -28,8 +61,10 @@ const ViewTitleBar = ({
                             >
                                 {displayName}
                             </span>
-                            {/* Todo: Re-add starring functionality */}
-                            <StarDashboardButton />
+                            <StarDashboardButton
+                                starred={starred}
+                                onClick={onToggleStarredDashboard}
+                            />
                         </div>
                     </div>
                     <div className={classes.detailWrap}>
@@ -38,7 +73,7 @@ const ViewTitleBar = ({
                             <div className={classes.groupedLastUpdated}>
                                 {<LastUpdatedTag id={id} />}
                             </div>
-                            <ActionsBar />
+                            <ActionsBar starred={starred} onToggleStarredDashboard={onToggleStarredDashboard} />
                         </div>
                     </div>
                 </div>
@@ -59,7 +94,9 @@ ViewTitleBar.propTypes = {
     displayDescription: PropTypes.string,
     displayName: PropTypes.string,
     id: PropTypes.string,
+    setDashboardStarred: PropTypes.func,
     showDescription: PropTypes.bool,
+    starred: PropTypes.bool,
 }
 
 const mapStateToProps = (state) => {
@@ -67,8 +104,13 @@ const mapStateToProps = (state) => {
 
     return {
         ...dashboard,
+        starred: dashboard.id
+            ? sGetDashboardStarred(state, dashboard.id)
+            : false,
         showDescription: sGetShowDescription(state),
     }
 }
 
-export default connect(mapStateToProps)(ViewTitleBar)
+export default connect(mapStateToProps, {
+    setDashboardStarred: acSetDashboardStarred,
+})(ViewTitleBar)
